Infantry AI, various tweaks
Infantry now move in zig-zags towards the hill. Has a small bug at the
moment that causes some waypoints to be skipped (will fix it later).

Changed AI to have an Update function so they can think as fast as they
move. Tick function is for more complicated calculations that don't need
to happen as often. This change was made because the AI was walking too
far past waypoints.

Some various tweak.s