Ships (continued) and Animate Sprites
Ships now spawn Infantry over a short period of time and then leave the
area (they appear transparent when there are no infantry on board).

More work on AnimatedSprites, including importing of dynamic animated
sprites from xml files (assuming they're in the style my exporter is).
Still needs node support but that won't be required for any animations
planned for this project. The only Sprite Animations planned for this
project are the Row Boats and that will just be a simple paddling
animation.